backend/app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
pgvector==0.3.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
ultralytics==8.3.0
sentence-transformers==3.1.1
pillow==10.4.0
mangum==0.17.0
reportlab==4.0.7
matplotlib==3.8.2
httpx==0.27.0
aiosmtplib==3.0.1
